removed centralized caching with metaclass (to keep it as history)
