+++ theano/gof/tests/test_vm.py
@@ -386,6 +386,9 @@ if run_memory_usage_tests:
 
 
 class RunOnce(theano.Op):
+
+    __props__ = ("nb_run",)
+
     def __init__(self):
         self.nb_run = 0
