+++ theano/gpuarray/dnn.py
@@ -13,7 +13,6 @@ from theano.scalar import as_scalar, constant, Log, get_scalar_type
 from theano.tensor import as_tensor_variable
 from theano.gradient import DisconnectedType, grad_not_implemented
 from theano.gof import Optimizer, local_optimizer, COp
-from theano.gof.utils import CacheClass
 from theano.gof.cmodule import GCC_compiler
 from theano.gof.type import CDataType, Generic
 from theano.compile import optdb
@@ -349,7 +348,7 @@ def version(raises=True):
 version.v = None
 
 
-class GpuDnnConvDesc(CacheClass, COp):
+class GpuDnnConvDesc(COp):
 
     """
     This Op builds a convolution descriptor for use in the other convolution
@@ -640,7 +639,7 @@ class GpuDnnConv(DnnBase):
         return [shape[2]]
 
 
-class GpuDnnConvGradW(CacheClass, DnnBase):
+class GpuDnnConvGradW(DnnBase):
 
     """
     The convolution gradient with respect to the weights.
@@ -771,7 +770,7 @@ class GpuDnnConvGradW(CacheClass, DnnBase):
         return [shape[2]]
 
 
-class GpuDnnConvGradI(CacheClass, DnnBase):
+class GpuDnnConvGradI(DnnBase):
     """
     The convolution gradient with respect to the inputs.
 
@@ -1138,8 +1137,9 @@ def dnn_gradweight(img, topgrad, kerns_shp, border_mode='valid',
     precision = get_precision(precision, [img, topgrad])
 
     desc = GpuDnnConvDesc(border_mode=border_mode, subsample=subsample,
-                          conv_mode=conv_mode, precision=precision)(kerns_shp)
-    out = GpuAllocEmpty(dtype=img.dtype, context_name=ctx_name)(*kerns_shp)
+                             conv_mode=conv_mode, precision=precision)(
+                                 kerns_shp)
+    out = GpuAllocEmpty(ctx_name, dtype=img.dtype)(*kerns_shp)
     return GpuDnnConvGradW()(img, topgrad, out, desc)
 
 
@@ -1150,7 +1150,6 @@ def dnn_gradweight3d(img, topgrad, kerns_shp, border_mode='valid',
     """
     return dnn_gradweight(img, topgrad, kerns_shp, border_mode,
                           subsample, conv_mode, precision)
-
 
 def dnn_gradinput(kerns, topgrad, img_shp, border_mode='valid',
                   subsample=(1, 1), conv_mode='conv', precision=None):
@@ -1166,8 +1165,9 @@ def dnn_gradinput(kerns, topgrad, img_shp, border_mode='valid',
     precision = get_precision(precision, [kerns, topgrad])
 
     desc = GpuDnnConvDesc(border_mode=border_mode, subsample=subsample,
-                          conv_mode=conv_mode, precision=precision)(kerns.shape)
-    out = GpuAllocEmpty(dtype=kerns.dtype, context_name=ctx_name)(*img_shp)
+                             conv_mode=conv_mode, precision=precision)(
+                                 kerns.shape)
+    out = GpuAllocEmpty(ctx_name, kerns.dtype)(*img_shp)
     return GpuDnnConvGradI()(kerns, topgrad, out, desc)
 
 
@@ -1178,7 +1178,6 @@ def dnn_gradinput3d(kerns, topgrad, img_shp, border_mode='valid',
     """
     return dnn_gradinput(kerns, topgrad, img_shp, border_mode, subsample,
                          conv_mode, precision)
-
 
 class GpuDnnPoolDesc(Op):
