+++ theano/tensor/var.py
@@ -12,7 +12,8 @@ import theano
 from theano.compat import PY3
 from theano.scalar import ComplexError, IntegerDivisionError
 from theano.gof import Constant, Variable
-from theano.gof.utils import hashtype, hash_from_ndarray
+from theano.gof.utils import hashtype
+from theano.tensor.utils import hash_from_ndarray
 from theano.tensor.type import TensorType
 from theano.configparser import config
